bld: makefile for building rt library using Clang
With GCC v4.9 (distributed by TI as MSP430 GCC toolchain), the
mementos does build and link with app using GCC, but the restore
to checkpoint does not seem to work. Linking the library bytecode
built with Clang does work. Mementos already does it, so we do
the same for DINO with this commit.