libdino: remove the dummying of dino iface from dino header
It doesn't make much sense to do it from within a DINO header because
then, the dependency to DINO is necessary (when the point of dummying
is to make the app buildable without the DINO dependency). So,
until we find a better solution, we replicate the dummying out code
in every app.